--- src/main/java/frc/robot/subsystems/drivetrain/commands/Drive.java
@@ -1,28 +1,0 @@
-package frc.robot.subsystems.drivetrain.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-
-public class Drive extends CommandBase {
-
-
-    @Override
-    public void initialize() {
-
-    }
-
-    @Override
-    public void execute() {
-
-    }
-
-    @Override
-    public void end(boolean interrupted) {
-        end(interrupted);
-    }
-
-    @Override
-    public boolean isFinished() {
-        return isFinished();
-    }
-}
-
