+++ src/main/java/frc/robot/subsystems/drivetrain/commands/Rotate.java
@@ -1,0 +1,37 @@
+package frc.robot.subsystems.drivetrain.commands;
+
+import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.drivetrain.SwerveDrive;
+import frc.robot.subsystems.drivetrain.SwerveModule;
+
+public class Rotate extends CommandBase {
+    private final SwerveModule swerveModule;
+    private final SwerveDrive swerveDrive;
+
+    public Rotate(SwerveModule swerveModule, SwerveDrive swerveDrive){
+        this.swerveDrive = swerveDrive;
+        this.swerveModule = swerveModule;
+        addRequirements(swerveDrive);
+    }
+
+
+    @Override
+    public void initialize() {
+
+    }
+
+    @Override
+    public void execute() {
+
+    }
+
+    @Override
+    public void end(boolean interrupted) {
+       end(interrupted);
+    }
+
+    @Override
+    public boolean isFinished() {
+        return false;
+    }
+}
